% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_costum_data.r
\name{prepare_custom_data}
\alias{prepare_custom_data}
\title{Prepare custom data in LinkedOmics-like format}
\usage{
prepare_custom_data(data, data_type = c("Omics", "Clinical"))
}
\arguments{
\item{data}{A data.frame or matrix in LinkedOmics-like format}

\item{data_type}{Character string specifying the data type.
Supported values are "Omics" and "Clinical".}
}
\value{
A cleaned data.frame ready for downstream analyses.
}
\description{
Prepare custom data in LinkedOmics-like format
}
\seealso{
LinkedOmicsLikeData
}
